Fix issues with None gradient and variable namespace
First, the `tf.clip_by_norm` wasn't working as some gradients were None (as
variables stored for profile are initialized to a constant, their gradient is
0 which appears to be None in this implementation). Therefore, these values are
skipped.
The following process doesn't appear to be impacted by this, as the value will
be assigned to the variables that are indeed cliped.

The second was that `self._name` have been forgotten on some tf.variable_scope
calls.