Used to create two tf.Session
Due to a Python specifity,

```
    def f(sess=tf.Session())
```

will always call `tf.Session()` once, even if the function is always called with
a specific session (which was done in `memn2n_dialog.py`).

This was not the expected behavior in the code, as it creates a useless session
that will never be closed in the code.

It has been changed in `MemN2N` and `MemN2N-mtl` to

```
    def f(sess=None):
        ...
        session = sess if sess is not None else tf.Session()
        ...
```

Therefore, `tf.Session()` is called only if it really called with `None` (and
several session will be created if the function is called more than once).